0x08-recursion

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -19,6 +19,6 @@ return (-1);
 }
 else
 {
-return (sqrt(n));
+return (sqrtf(n));
 }
 }
